`PythonFunction` causes GitHub Action failure
Locally testing our GitHub Action via `act` fails. It appears that
perhaps this is due to an attempt by `PythonFunction` to build a lambda
function using a Docker container (under the hood), but doing so within
the context of the GitHub Action Docker container causes a problem.

Using the more general `Function` in place of `PythonFunction` resolves
the failure.