Conditionally install node and cdk in virtualenv
To speedup local development, we want to avoid reinstalling node and cdk
when the correct versions are already installed within the virtualenv.
However, we also want to make sure we properly test for their existence
within the virtualenv, and not outside of it, and also make sure the
correct versions are installed.